requirements.txt
amqp==2.6.1
billiard==3.6.3.0
celery==4.4.7
certifi==2020.6.20
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
Django==2.2.15
django-environ==0.4.5
django-rest-auth==0.9.5
django-rest-swagger==2.1.2
djangorestframework==3.11.1
djangorestframework-jwt==1.11.0
gunicorn==20.0.4
idna==2.10
importlib-metadata==1.7.0
itypes==1.2.0
Jinja2==2.11.2
kombu==4.6.11
MarkupSafe==1.1.1
mysqlclient==2.0.1
openapi-codec==1.3.2
PyJWT==1.7.1
PyMySQL==0.10.0
pytz==2020.1
redis==3.5.3
requests==2.24.0
selenium==3.141.0
simplejson==3.17.2
six==1.15.0
sqlparse==0.3.1
uritemplate==3.0.1
urllib3==1.25.10
vine==1.3.0
waitress==1.4.4
zipp==1.2.0
